+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Moved static base URL in the ENV config
 - Update Cere Wallet SDK to add CERE ERC20 tokens support
+- Log and store session events
 
 ### v0.1.2